web/requirements.txt
click==8.1.3
fastapi==0.87.0
Jinja2==3.1.2
Pillow==9.3.0
python-multipart==0.0.5
requests==2.28.1
uvicorn==0.20.0